requirements.txt
langchain
openai
python-dotenv
fastapi
uvicorn
streamlit
langchain_community
python-multipart